+++ kubernetes/types.go
@@ -25,7 +25,6 @@ import (
 	batchv1 "k8s.io/api/batch/v1"
 	v1 "k8s.io/api/core/v1"
 	networkingv1 "k8s.io/api/networking/v1"
-	policyv1beta1 "k8s.io/api/policy/v1beta1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -109,9 +108,6 @@ type ClusterRole = rbacv1.ClusterRole
 
 // ClusterRoleBinding data
 type ClusterRoleBinding = rbacv1.ClusterRoleBinding
-
-// PodSecurityPolicy data
-type PodSecurityPolicy = policyv1beta1.PodSecurityPolicy
 
 // NetworkPolicy data
 type NetworkPolicy = networkingv1.NetworkPolicy